Revert "feat(lib): add time.rs for monotonically increasig time (#244)" (#247)
This reverts commit f3055e5d735ced3ada4210f791682a4afe9b2dab.

+++ awkernel_lib/src/delay.rs
@@ -33,7 +33,7 @@ pub trait Delay {
     fn uptime() -> u64;
 
     /// This function returns uptime in nanoseconds.
-    fn uptime_nano() -> u64;
+    fn uptime_nano() -> u128;
 
     /// Return CPU cycle counter.
     fn cpu_counter() -> u64;
@@ -119,7 +119,7 @@ pub fn uptime() -> u64 {
 /// log::info!("{} [ns]", end - start);
 /// ```
 #[inline(always)]
-pub fn uptime_nano() -> u64 {
+pub fn uptime_nano() -> u128 {
     ArchImpl::uptime_nano()
 }
